signal: add `Clear`, `Stop`, make `Kernel` garbage collect
When using `Clear` and `Stop`, trying to route `to_gc` channels
within the `someday` writer via signals such that `Kernel` isn't
the one dropping within `Apply`, but it just sends it `to_gc`...

...is very very hard and complicated to wire up.

Since `Kernel` isn't necessarily real-time anyway, just make it
clean up when something like `Stop` comes around.

This means if the queue is very large, `Kernel` will be responsible
for `.clear()`'ing that twice - 1 for the writer, 1 for the reader.