handle errors with callbacks, setup `Kernel` routing handlers
The `ErrorBehavior` is now `ErrorCallback` inside `Callbacks`.
`Audio/Decode` will report to `Kernel` when errors happen, and
`Kernel` must determine what to do since it is the only owner
of the (potential) callback functions.

If `ErrorCallback::Continue` is selected, `Audio/Decode` will not
wait on a `Kernel` message and will continue as normal, otherwise
they will wait until `Kernel` has handled the error and responded.