remove `Pool`, clear buffers by draining to `Gc`
It takes 0.00006~ seconds to `.pop()` and send 8,000 `AudioBuffer`'s
over a channel, which is fast enough where when `Decode` needs
to clear its `VecDeque`, it can just do that instead of the
swap with `Pool`.

This does mean the clear time scales with `N` where `N` is the
amount of things we need to pop and send instead of a constant
1 send, 1 recv from `Pool`, but it's worth it since `Pool` no
longer needs to be a thing.